Move Credential calls and types to CredentialsMixin
