+++ gvm/protocols/gmpv214/types.py
@@ -23,29 +23,21 @@ from gvm.errors import InvalidArgument
 
 from gvm.protocols.gmpv208.types import (
     AggregateStatistic,
-    CredentialFormat,
-    CredentialType,
     EntityType,
     FeedType,
     FilterType,
     HostsOrdering,
     PermissionSubjectType,
-    SnmpAuthAlgorithm,
-    SnmpPrivacyAlgorithm,
     SortOrder,
     TicketStatus,
     TimeUnit,
     UserAuthType,
     get_aggregate_statistic_from_string,
-    get_credential_format_from_string,
-    get_credential_type_from_string,
     get_entity_type_from_string,
     get_feed_type_from_string,
     get_filter_type_from_string,
     get_hosts_ordering_from_string,
     get_permission_subject_type_from_string,
-    get_snmp_auth_algorithm_from_string,
-    get_snmp_privacy_algorithm_from_string,
     get_sort_order_from_string,
     get_ticket_status_from_string,
     get_time_unit_from_string,
@@ -55,8 +47,6 @@ from gvm.protocols.gmpv208.types import (
 
 __all__ = [
     "AggregateStatistic",
-    "CredentialFormat",
-    "CredentialType",
     "EntityType",
     "FeedType",
     "FilterType",
@@ -64,15 +54,11 @@ __all__ = [
     "PermissionSubjectType",
     "ScannerType",
     "SeverityLevel",
-    "SnmpAuthAlgorithm",
-    "SnmpPrivacyAlgorithm",
     "SortOrder",
     "TicketStatus",
     "TimeUnit",
     "UserAuthType",
     "get_aggregate_statistic_from_string",
-    "get_credential_format_from_string",
-    "get_credential_type_from_string",
     "get_entity_type_from_string",
     "get_feed_type_from_string",
     "get_filter_type_from_string",
@@ -80,15 +66,13 @@ __all__ = [
     "get_permission_subject_type_from_string",
     "get_scanner_type_from_string",
     "get_severity_level_from_string",
-    "get_snmp_auth_algorithm_from_string",
-    "get_snmp_privacy_algorithm_from_string",
     "get_sort_order_from_string",
     "get_ticket_status_from_string",
     "get_time_unit_from_string",
     "get_user_auth_type_from_string",
 ]
 
-
+# move this: Severity Level !!!
 class SeverityLevel(Enum):
     """Enum for severity levels"""
